update23

+++ quanlykhodl/quanlykhodl/ViewModel/DeliverynoteGetAll.cs
@@ -18,7 +18,7 @@
         public string? code { get; set; }
         public int? Tax { get; set; }
         public int? TotalProduct { get; set; }
-        public int? TotalQuantity { get; set; }
+        public long? TotalQuantity { get; set; }
         public bool? isPercentage { get; set; }
         public List<productImportformAndDeliveerrynote>? products { get; set; }
     }
